+++ apps/learn-card-app/src/components/familyCMS/FamilyBoostPreview/FamilyBoostMembersList/FamilyBoostMemberListItem.tsx
@@ -45,10 +45,12 @@ export const FamilyBoostMembersListItem: React.FC<{
         );
     };
 
+    const displayName = managedProfile?.displayName || user?.displayName || user?.profileId;
+
     return (
-        <div
+        <button
             onClick={handleMemberActionMenu}
-            className="flex items-center justify-between w-full bg-white border-t-grayscale-100 border-t-solid border-t-[2px] pb-[12px] pt-[12px]"
+            className="flex items-center justify-between w-full bg-white border-t-grayscale-100 border-solid border-t-[2px] pb-[12px] pt-[12px]"
         >
             <div className="flex items-center justify-start w-full">
                 <div className="flex items-center justify-start">
@@ -67,16 +69,14 @@ export const FamilyBoostMembersListItem: React.FC<{
                     {currentUserIsActiveUser && (
                         <p className="text-emerald-700 text-xs font-semibold">Active User</p>
                     )}
-                    <p className="text-grayscale-900 font-normal font-poppins">
-                        {user?.displayName || user?.profileId}
-                    </p>
+                    <p className="text-grayscale-900 font-normal font-poppins">{displayName}</p>
                 </div>
             </div>
-            <button className="flex items-center justify-center text-grayscale-600 font-poppins text-sm">
+            <div className="flex items-center justify-center text-grayscale-600 font-poppins text-sm">
                 {familyTitles?.singular}
                 <SlimCaretRight className="text-grayscale-400 w-[20px] h-auto" />
-            </button>
-        </div>
+            </div>
+        </button>
     );
 };