+++ packages/plugins/did-web-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # did-web-plugin
 
+## 1.1.7
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.4.7
+
 ## 1.1.6
 
 ### Patch Changes